Conflicts:
	src/networklayer/ipv6/IPv6Datagram.cc

+++ src/linklayer/ext/ExtInterface.h
@@ -41,17 +41,17 @@ class ExtInterface : public cSimpleModule
   protected:
     bool connected;
     uint8 buffer[1<<16];
-	const char *device;
+    const char *device;
 
     InterfaceEntry *interfaceEntry;  // points into RoutingTable
 
     // statistics
-    uint64 numSent;
-    uint64 numRcvd;
-    uint64 numDropped;
+    int numSent;
+    int numRcvd;
+    int numDropped;
 
-	// access to real network interface via Scheduler class:
-	cSocketRTScheduler *rtScheduler;
+    // access to real network interface via Scheduler class:
+    cSocketRTScheduler *rtScheduler;
 
     InterfaceEntry *registerInterface();
     void displayBusy();
@@ -69,7 +69,7 @@ class ExtInterface : public cSimpleModule
     virtual void initialize(int stage);
     virtual void handleMessage(cMessage *msg);
 
-	virtual void finish();
+    virtual void finish();
 };
 
 #endif
